Fixed the issue identified by this test.
this branch should not be merged with master, ever. instead, i'll
implement this fix into the join branch, then work through the merges
from there as necessary. this branch should probably be abandoned
altogether for a new branch based on the updated versions of
join and condition2. i could merge them to update this one. not
sure which is correct