Changed if to while. Needs merging.
honestly this commit is a mess. i made a billion changes to make testing
easier, and it makes the code a mess. i'll fix the problem on the comms
branch and then merge with this branch. that should remove most of the
mess. if not, we aren't gonna submit anything on this branch anyway;
all the changes should have been made to self tests.